LPS-162847 Move useId to core/hooks
