+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/app/components/fragment-configuration-fields/TargetCollectionsField.js
@@ -18,12 +18,12 @@ import ClayForm, {ClayCheckbox} from '@clayui/form';
 import classNames from 'classnames';
 import React, {useState} from 'react';
 
+import {useId} from '../../../core/hooks/useId';
 import {LAYOUT_DATA_ITEM_TYPES} from '../../config/constants/layoutDataItemTypes';
 import {useHoverItem} from '../../contexts/ControlsContext';
 import {useSelectorCallback} from '../../contexts/StoreContext';
 import isEmptyArray from '../../utils/isEmptyArray';
 import {isLayoutDataItemDeleted} from '../../utils/isLayoutDataItemDeleted';
-import {useId} from '../../utils/useId';
 
 export function selectConfiguredCollectionDisplays(state) {
 	return Object.values(state.layoutData.items).filter(
