+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/common/components/ColorPalette.js
@@ -18,7 +18,7 @@ import PropTypes from 'prop-types';
 import React, {useMemo} from 'react';
 
 import {config} from '../../app/config/index';
-import {useId} from '../../app/utils/useId';
+import {useId} from '../../core/hooks/useId';
 
 export default function ColorPalette({
 	label,
