+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/common/components/ImageSelector.js
@@ -20,7 +20,7 @@ import React from 'react';
 import {VIEWPORT_SIZES} from '../../app/config/constants/viewportSizes';
 import {useSelector} from '../../app/contexts/StoreContext';
 import {selectPageContents} from '../../app/selectors/selectPageContents';
-import {useId} from '../../app/utils/useId';
+import {useId} from '../../core/hooks/useId';
 import {openImageSelector} from '../../core/openImageSelector';
 
 export function ImageSelector({
