+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/common/components/ImageSelectorSize.js
@@ -24,7 +24,7 @@ import selectLanguageId from '../../app/selectors/selectLanguageId';
 import ImageService from '../../app/services/ImageService';
 import isMapped from '../../app/utils/editable-value/isMapped';
 import resolveEditableValue from '../../app/utils/editable-value/resolveEditableValue';
-import {useId} from '../../app/utils/useId';
+import {useId} from '../../core/hooks/useId';
 
 export const DEFAULT_IMAGE_SIZE_ID = 'auto';
