+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/common/components/MappingFieldSelector.js
@@ -18,7 +18,7 @@ import React from 'react';
 
 import {EDITABLE_TYPES} from '../../app/config/constants/editableTypes';
 import getSelectedField from '../../app/utils/getSelectedField';
-import {useId} from '../../app/utils/useId';
+import {useId} from '../../core/hooks/useId';
 
 const UNMAPPED_OPTION = {
 	label: `-- ${Liferay.Language.get('unmapped')} --`,
