+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/plugins/browser/components/page-structure/components/item-configuration-panels/ImageSourcePanel.js
@@ -36,10 +36,10 @@ import isMappedToInfoItem from '../../../../../../app/utils/editable-value/isMap
 import {getEditableLocalizedValue} from '../../../../../../app/utils/getEditableLocalizedValue';
 import {setIn} from '../../../../../../app/utils/setIn';
 import {updateIn} from '../../../../../../app/utils/updateIn';
-import {useId} from '../../../../../../app/utils/useId';
 import {ImageSelector} from '../../../../../../common/components/ImageSelector';
 import {ImageSelectorDescription} from '../../../../../../common/components/ImageSelectorDescription';
 import {ImageSelectorSize} from '../../../../../../common/components/ImageSelectorSize';
+import {useId} from '../../../../../../core/hooks/useId';
 import {getEditableItemPropTypes} from '../../../../../../prop-types/index';
 import {MappingPanel} from './MappingPanel';
