+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/plugins/browser/components/page-structure/components/item-configuration-panels/collection-general-panel/StyleDisplaySelector.js
@@ -17,7 +17,7 @@ import PropTypes from 'prop-types';
 import React, {useEffect, useState} from 'react';
 
 import InfoItemService from '../../../../../../../app/services/InfoItemService';
-import {useId} from '../../../../../../../app/utils/useId';
+import {useId} from '../../../../../../../core/hooks/useId';
 
 const LIST_STYLE_GRID = '';
